modularise for ease of use, include samples loop as part of class to harmonise with diff analysis, remove diff analysis for now ahead of testing
